+++ roulier/carriers/dpd/dpd_transport.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
 """Implement dpdWS."""
 import requests
-import email.parser
 from lxml import objectify, etree
 from jinja2 import Environment, PackageLoader
 from roulier.transport import Transport
 from roulier.ws_tools import remove_empty_tags
+from roulier.exception import CarrierError
+
 import logging
 
 log = logging.getLogger(__name__)
@@ -15,8 +16,6 @@ class DpdTransport(Transport):
     """Implement Dpd WS communication."""
 
     DPD_WS = "https://e-station.cargonet.software/dpd-eprintwebservice/eprintwebservice.asmx"
-    STATUS_SUCCES = "success"
-    STATUS_ERROR = "error"
 
     def send(self, payload):
         """Call this function.
@@ -26,11 +25,8 @@ class DpdTransport(Transport):
             payload.header : auth
         Return:
             {
-                status: STATUS_SUCCES or STATUS_ERROR, (string)
-                message: more info about status of result (lxml)
                 response: (Requests.response)
-                payload: usefull payload (if success) (xml as string)
-
+                body: XML response (without soap)
             }
         """
         body = payload['body']
@@ -65,21 +61,14 @@ class DpdTransport(Transport):
         """Handle reponse in case of ERROR 500 type."""
         log.warning('Dpd error 500')
         obj = objectify.fromstring(response.content)
-        return {
+        errors = [{
             "id": obj.xpath('//faultcode')[0],
-            "status": self.STATUS_ERROR,
             "message": obj.xpath('//faultstring')[0],
-            "response": response,
-            "payload": None
-        }
+        }]
+        raise CarrierError(response, errors)
 
     def handle_200(self, response):
-        """
-        Handle response type 200 (success).
-
-        It still can be a success or a failure.
-        """
-
+        """Handle response type 200 (success)."""
         def extract_soap(response_xml):
             obj = objectify.fromstring(response_xml)
             return obj.Body.getchildren()[0]
@@ -87,22 +76,18 @@ class DpdTransport(Transport):
         body = extract_soap(response.content)
         body_xml = etree.tostring(body)
         return {
-            "status": "ok",
-            "message": "",
-            "payload": body_xml,
+            "body": body_xml,
             "response": response,
         }
 
     def handle_response(self, response):
         """Handle response of webservice."""
-        if response.status_code == 500:
-            return self.handle_500(response)
-        elif response.status_code == 200:
+        if response.status_code == 200:
             return self.handle_200(response)
+        elif response.status_code == 500:
+            return self.handle_500(response)
         else:
-            return {
-                "status": "error",
-                "message": "Unexpected status code from server",
-                "response": response
-            }
-
+            raise CarrierError(response, [{
+                'id': None,
+                'message': "Unexpected status code from server",
+            }])