+++ roulier/carriers/laposte/laposte.py
@@ -22,15 +22,11 @@ class Laposte(Carrier):
         """Run an action with data against Laposte WS."""
         request = self.encoder.encode(data, action)
         response = self.ws.send(request)
-        if response.get('message') and response['message']['exception']:
-            return {
-                'status': 'error',
-                'messages': self.ws.exception_handling(
-                    response['message']['message']),
-                'response': response['response'],
-            }
-        parts = self.ws.get_parts(response['response'])
-        return self.decoder.decode(response, parts)
+        return self.decoder.decode(
+            response['body'],
+            response['parts'],
+            request['output_format']
+        )
 
     # shortcuts
     def get_label(self, data):